llm-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
llama-cpp-python==0.2.90

